requirements.txt
beautifulsoup4
requests
newspaper3k
pyap
